zest: FingerprintStyle: Make the feature work
Re-style the activity to slightly match One UI more, and make the feature work. Needs system modifications but I'm happy with this.

We still have to check if we are 'provisioned' so existing animation doesn't get carried over to the next data wipe as well as code cleanup but this is functional now.

Signed-off-by: John Vincent <git@tenseventyseven.cf>