fresh: SoftwareUpdate: Switch to using JobScheduler for update checking
So we don't have to worry about firing these checks even if the user doesn't have an internet connection. Let's also separate the checks done on service and activity so we can do whatever we like on those, since we use callbacks for it anyway.

While we're at it, move a lot of stuff from UpdateUtils to appropriate classes.

Signed-off-by: John Vincent <git@tenseventyseven.cf>