fresh: SoftwareUpdate: Make software update downloads actually work
This adds a download service based on work by Rahul Patel for AuroraDroid (see OSS licenses), which allows the user to queue, pause, resume, and cancel downloads from the notification. I've added a bit of 'improvements' as download notifications from AuroraDroid are normally dismissible (regardless of state) as well as adapting it for our use.

There's a couple more things to work on for the activity (i.e. catch more instances of the device losing a connection, cancellations, etc), but this is fully functional at this point.

Signed-off-by: John Vincent <git@tenseventyseven.cf>