fresh: Initial work on migrating to new UI library
This is an initial work on moving to OneUIProject/oneui-core and OneUIProject/oneui-design. Lots of missing stuff on the new library and app is currently not usable at this state, hence this being on a separate branch. Changes from the main 'fresh-4.x' branch will also be merged here.

Merge once we are finished with migration. Pull requests are welcome.

Signed-off-by: John Vincent <git@tenseventyseven.cf>