fresh: SoftwareUpdate: Fix update post-download in the background
There's a tiny bug (that turns out, needed huge changes to fix lmao) where package hash verification won't proceed if the app is closed in the background, causing the AvailableActivity screen to won't proceed with installation.

Fix it by revamping how the service updates the notification, as well as how it finishes the download.

Signed-off-by: John Vincent <git@tenseventyseven.cf>