fresh: ScreenResolution: Make it possible to change display resolution while keeping aspect ratio
Commonize the resolution override and add ratio calculations for it, based on the set aspect ratio in Zest. This way, we can allow the user to change the screen resolution, while still keeping their preferred display aspect ratio and removing the need to reset it everytime the user changes their resolution. This should make these settings more intuitive.

Signed-off-by: John Vincent <git@tenseventyseven.cf>