zest: Drop storage permission checks
For an unknown reason the permission checks fail and the user will always be prompted to grant storage access, regardless of whether the permissions have already been granted.

As FSS is a system app and storage permissions are marked as fixed, we can safely drop the checks and still expect storage access to be granted.

Signed-off-by: Louis Poier <lpoier7@proton.me>